fix: resolve TypeScript errors and add test mocks
